split unstable_derive test
