+++ tests/vanilla/store.test.tsx
@@ -1,7 +1,7 @@
 import { waitFor } from '@testing-library/dom'
 import { assert, describe, expect, it, vi } from 'vitest'
 import { atom, createStore } from 'jotai/vanilla'
-import type { Atom, ExtractAtomValue, Getter } from 'jotai/vanilla'
+import type { Getter } from 'jotai/vanilla'
 
 it('should not fire on subscribe', async () => {
   const store = createStore()
@@ -571,346 +571,4 @@ it('Unmount an atom that is no longer dependent within a derived atom (#2658)', 
   store.sub(derivedAtom, () => {})
   store.set(condAtom, false)
   expect(onUnmount).toHaveBeenCalledTimes(1)
-})
-
-describe('unstable_derive for scoping atoms', () => {
-  /**
-   * a
-   * S1[a]: a1
-   */
-  it('primitive atom', async () => {
-    const a = atom('a')
-    a.onMount = (setSelf) => setSelf((v) => v + ':mounted')
-    const scopedAtoms = new Set<Atom<unknown>>([a])
-
-    const store = createStore()
-    const derivedStore = store.unstable_derive((getAtomState) => {
-      const scopedAtomStateMap = new WeakMap()
-      return [
-        (atom, originAtomState) => {
-          if (scopedAtoms.has(atom)) {
-            let atomState = scopedAtomStateMap.get(atom)
-            if (!atomState) {
-              atomState = { d: new Map(), p: new Set(), n: 0 }
-              scopedAtomStateMap.set(atom, atomState)
-            }
-            return atomState
-          }
-          return getAtomState(atom, originAtomState)
-        },
-      ]
-    })
-
-    expect(store.get(a)).toBe('a')
-    expect(derivedStore.get(a)).toBe('a')
-
-    derivedStore.sub(a, vi.fn())
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a')
-    expect(derivedStore.get(a)).toBe('a:mounted')
-
-    derivedStore.set(a, (v) => v + ':updated')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a')
-    expect(derivedStore.get(a)).toBe('a:mounted:updated')
-  })
-
-  /**
-   * a, b, c(a + b)
-   * S1[a]: a1, b0, c0(a1 + b0)
-   */
-  it('derived atom (scoping primitive)', async () => {
-    const a = atom('a')
-    const b = atom('b')
-    const c = atom((get) => get(a) + get(b))
-    const scopedAtoms = new Set<Atom<unknown>>([a])
-
-    const store = createStore()
-    const derivedStore = store.unstable_derive((getAtomState) => {
-      type AnyAtom = Parameters<typeof getAtomState>[0]
-      type AtomState = ReturnType<typeof getAtomState>
-      const scopedAtomStateMap = new WeakMap<AnyAtom, AtomState>()
-      const scopedInvertAtomStateMap = new WeakMap<AtomState, AnyAtom>()
-      const copyMounted = (
-        mounted: NonNullable<AtomState['m']>,
-      ): NonNullable<AtomState['m']> => ({
-        ...mounted,
-        d: new Set(mounted.d),
-        t: new Set(mounted.t),
-      })
-      const copyAtomState = (atomState: AtomState): AtomState => ({
-        ...atomState,
-        d: new Map(atomState.d),
-        p: new Set(atomState.p),
-        ...('m' in atomState ? { m: copyMounted(atomState.m) } : {}),
-      })
-      return [
-        (atom, originAtomState) => {
-          type TheAtomState = ReturnType<
-            typeof getAtomState<ExtractAtomValue<typeof atom>>
-          >
-          let atomState = scopedAtomStateMap.get(atom)
-          if (atomState) {
-            return atomState as TheAtomState
-          }
-          if (
-            scopedInvertAtomStateMap.has(originAtomState as never) ||
-            scopedAtoms.has(atom)
-          ) {
-            atomState = { d: new Map(), p: new Set(), n: 0 }
-            scopedAtomStateMap.set(atom, atomState)
-            scopedInvertAtomStateMap.set(atomState, atom)
-            return atomState as TheAtomState
-          }
-          const originalAtomState = getAtomState(atom, originAtomState)
-          if (
-            Array.from(originalAtomState.d).some(([a]) =>
-              scopedAtomStateMap.has(a),
-            )
-          ) {
-            atomState = copyAtomState(originalAtomState)
-            scopedAtomStateMap.set(atom, atomState)
-            scopedInvertAtomStateMap.set(atomState, atom)
-            return atomState as TheAtomState
-          }
-          return originalAtomState
-        },
-      ]
-    })
-
-    expect(store.get(c)).toBe('ab')
-    expect(derivedStore.get(c)).toBe('ab')
-
-    derivedStore.set(a, 'a2')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(c)).toBe('ab')
-    expect(derivedStore.get(c)).toBe('a2b')
-
-    derivedStore.sub(c, vi.fn())
-    derivedStore.set(b, 'b2')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(c)).toBe('ab2')
-    expect(derivedStore.get(c)).toBe('a2b2')
-  })
-
-  /**
-   * a, b(a)
-   * S1[b]: a0, b1(a1)
-   */
-  it('derived atom (scoping derived)', async () => {
-    const a = atom('a')
-    const b = atom(
-      (get) => get(a),
-      (_get, set, v: string) => {
-        set(a, v)
-      },
-    )
-    const scopedAtoms = new Set<Atom<unknown>>([b])
-
-    const store = createStore()
-    const derivedStore = store.unstable_derive((getAtomState) => {
-      const scopedAtomStateMap = new WeakMap()
-      const scopedAtomStateSet = new WeakSet()
-      return [
-        (atom, originAtomState) => {
-          if (
-            scopedAtomStateSet.has(originAtomState as never) ||
-            scopedAtoms.has(atom)
-          ) {
-            let atomState = scopedAtomStateMap.get(atom)
-            if (!atomState) {
-              atomState = { d: new Map(), p: new Set(), n: 0 }
-              scopedAtomStateMap.set(atom, atomState)
-              scopedAtomStateSet.add(atomState)
-            }
-            return atomState
-          }
-          return getAtomState(atom, originAtomState)
-        },
-      ]
-    })
-
-    expect(store.get(a)).toBe('a')
-    expect(store.get(b)).toBe('a')
-    expect(derivedStore.get(a)).toBe('a')
-    expect(derivedStore.get(b)).toBe('a')
-
-    store.set(a, 'a2')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a2')
-    expect(store.get(b)).toBe('a2')
-    expect(derivedStore.get(a)).toBe('a2')
-    expect(derivedStore.get(b)).toBe('a')
-
-    store.set(b, 'a3')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a3')
-    expect(store.get(b)).toBe('a3')
-    expect(derivedStore.get(a)).toBe('a3')
-    expect(derivedStore.get(b)).toBe('a')
-
-    derivedStore.set(a, 'a4')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a4')
-    expect(store.get(b)).toBe('a4')
-    expect(derivedStore.get(a)).toBe('a4')
-    expect(derivedStore.get(b)).toBe('a')
-
-    derivedStore.set(b, 'a5')
-    await new Promise((resolve) => setTimeout(resolve))
-    expect(store.get(a)).toBe('a4')
-    expect(store.get(b)).toBe('a4')
-    expect(derivedStore.get(a)).toBe('a4')
-    expect(derivedStore.get(b)).toBe('a5')
-  })
-
-  /**
-   * a, b, c(a), d(c), e(d + b)
-   * S1[d]: a0, b0, c0(a0), d1(c1(a1)), e0(d1(c1(a1)) + b0)
-   */
-  it('derived atom (scoping derived chain)', async () => {
-    const a = atom('a')
-    const b = atom('b')
-    const c = atom(
-      (get) => get(a),
-      (_get, set, v: string) => set(a, v),
-    )
-    const d = atom(
-      (get) => get(c),
-      (_get, set, v: string) => set(c, v),
-    )
-    const e = atom(
-      (get) => get(d) + get(b),
-      (_get, set, av: string, bv: string) => {
-        set(d, av)
-        set(b, bv)
-      },
-    )
-    const scopedAtoms = new Set<Atom<unknown>>([d])
-
-    function makeStores() {
-      const baseStore = createStore()
-      const deriStore = baseStore.unstable_derive((getAtomState) => {
-        const scopedAtomStateMap = new WeakMap()
-        const scopedAtomStateSet = new WeakSet()
-        return [
-          (atom, originAtomState) => {
-            if (
-              scopedAtomStateSet.has(originAtomState as never) ||
-              scopedAtoms.has(atom)
-            ) {
-              let atomState = scopedAtomStateMap.get(atom)
-              if (!atomState) {
-                atomState = { d: new Map(), p: new Set(), n: 0 }
-                scopedAtomStateMap.set(atom, atomState)
-                scopedAtomStateSet.add(atomState)
-              }
-              return atomState
-            }
-            return getAtomState(atom, originAtomState)
-          },
-        ]
-      })
-      expect(getAtoms(baseStore)).toEqual(['a', 'b', 'a', 'a', 'ab'])
-      expect(getAtoms(deriStore)).toEqual(['a', 'b', 'a', 'a', 'ab'])
-      return { baseStore, deriStore }
-    }
-    type Store = ReturnType<typeof createStore>
-    function getAtoms(store: Store) {
-      return [
-        store.get(a),
-        store.get(b),
-        store.get(c),
-        store.get(d),
-        store.get(e),
-      ]
-    }
-
-    /**
-     * base[d]: a0, b0, c0(a0), d0(c0(a0)), e0(d0(c0(a0)) + b0)
-     * deri[d]: a0, b0, c0(a0), d1(c1(a1)), e0(d1(c1(a1)) + b0)
-     */
-    {
-      // UPDATE a0
-      // NOCHGE b0 and a1
-      const { baseStore, deriStore } = makeStores()
-      baseStore.set(a, '*')
-      expect(getAtoms(baseStore)).toEqual(['*', 'b', '*', '*', '*b'])
-      expect(getAtoms(deriStore)).toEqual(['*', 'b', '*', 'a', 'ab'])
-    }
-    {
-      // UPDATE b0
-      // NOCHGE a0 and a1
-      const { baseStore, deriStore } = makeStores()
-      baseStore.set(b, '*')
-      expect(getAtoms(baseStore)).toEqual(['a', '*', 'a', 'a', 'a*'])
-      expect(getAtoms(deriStore)).toEqual(['a', '*', 'a', 'a', 'a*'])
-    }
-    {
-      // UPDATE c0, c0 -> a0
-      // NOCHGE b0 and a1
-      const { baseStore, deriStore } = makeStores()
-      baseStore.set(c, '*')
-      expect(getAtoms(baseStore)).toEqual(['*', 'b', '*', '*', '*b'])
-      expect(getAtoms(deriStore)).toEqual(['*', 'b', '*', 'a', 'ab'])
-    }
-    {
-      // UPDATE d0, d0 -> c0 -> a0
-      // NOCHGE b0 and a1
-      const { baseStore, deriStore } = makeStores()
-      baseStore.set(d, '*')
-      expect(getAtoms(baseStore)).toEqual(['*', 'b', '*', '*', '*b'])
-      expect(getAtoms(deriStore)).toEqual(['*', 'b', '*', 'a', 'ab'])
-    }
-    {
-      // UPDATE e0, e0 -> d0 -> c0 -> a0
-      //             └--------------> b0
-      // NOCHGE a1
-      const { baseStore, deriStore } = makeStores()
-      baseStore.set(e, '*', '*')
-      expect(getAtoms(baseStore)).toEqual(['*', '*', '*', '*', '**'])
-      expect(getAtoms(deriStore)).toEqual(['*', '*', '*', 'a', 'a*'])
-    }
-    {
-      // UPDATE a0
-      // NOCHGE b0 and a1
-      const { baseStore, deriStore } = makeStores()
-      deriStore.set(a, '*')
-      expect(getAtoms(baseStore)).toEqual(['*', 'b', '*', '*', '*b'])
-      expect(getAtoms(deriStore)).toEqual(['*', 'b', '*', 'a', 'ab'])
-    }
-    {
-      // UPDATE b0
-      // NOCHGE a0 and a1
-      const { baseStore, deriStore } = makeStores()
-      deriStore.set(b, '*')
-      expect(getAtoms(baseStore)).toEqual(['a', '*', 'a', 'a', 'a*'])
-      expect(getAtoms(deriStore)).toEqual(['a', '*', 'a', 'a', 'a*'])
-    }
-    {
-      // UPDATE c0, c0 -> a0
-      // NOCHGE b0 and a1
-      const { baseStore, deriStore } = makeStores()
-      deriStore.set(c, '*')
-      expect(getAtoms(baseStore)).toEqual(['*', 'b', '*', '*', '*b'])
-      expect(getAtoms(deriStore)).toEqual(['*', 'b', '*', 'a', 'ab'])
-    }
-    {
-      // UPDATE d1, d1 -> c1 -> a1
-      // NOCHGE b0 and a0
-      const { baseStore, deriStore } = makeStores()
-      deriStore.set(d, '*')
-      expect(getAtoms(baseStore)).toEqual(['a', 'b', 'a', 'a', 'ab'])
-      expect(getAtoms(deriStore)).toEqual(['a', 'b', 'a', '*', '*b'])
-    }
-    {
-      // UPDATE e0, e0 -> d1 -> c1 -> a1
-      //             └--------------> b0
-      // NOCHGE a0
-      const { baseStore, deriStore } = makeStores()
-      deriStore.set(e, '*', '*')
-      expect(getAtoms(baseStore)).toEqual(['a', '*', 'a', 'a', 'a*'])
-      expect(getAtoms(deriStore)).toEqual(['a', '*', 'a', '*', '**'])
-    }
-  })
 })
